Add script to copy SSH keys from ks secrets
This script is intended to be run once after the initial setup
to copy SSH keys from ks secrets to the appropriate locations
in the user's SSH directory. It reads the SSH config file to
determine which keys need to be copied and handles the copying
and setting of appropriate permissions.